+++ packages/web3-providers/src/Uniswap/index.ts
@@ -1,7 +1,7 @@
 import { DataProvider } from '@masknet/public-api'
 import { TokenType } from '@masknet/web3-shared-base'
 import type { ChainId } from '@masknet/web3-shared-evm'
-import { isMirroredKeyword } from '../trending/helpers.js'
+import { isMirroredKeyword } from '../Trending/helpers.js'
 import type { TrendingAPI } from '../types/index.js'
 import * as BaseAPI from './base-api.js'
 import { BTC_FIRST_LEGER_DATE, getAllCoinsByKeyword, getPriceStats as getStats } from './base-api.js'
